Tidied Repo
Removed the older files for GP / mean / kernel and replaced them with the newer versions.

Tests have been removed at this stage as they were not fit for purpose. They shall be readded before merging with `master`.
