+++ gpybo/__init__.py
@@ -1,2 +1,1 @@
 from .gp import GP
-from .kernel import SquaredExponentialKernel
