+++ gpybo/gp.py
@@ -1,53 +1,64 @@
 from copy import deepcopy
-from typing import Any, Sequence
+from typing import Any, Tuple
 
+import numpy as np
 import torch
 import torch.nn as nn
 from torch import Tensor
 from torch.distributions.multivariate_normal import MultivariateNormal
 
-from .distributions import Normal
-from .kernel import Kernel
-from .mean import Mean, ZeroMean
-from .likelihood import GaussianLikelihood
-from .utils.early_stopping import EarlyStopping
-from .utils.shaping import to_tensor, uprank_two, unwrap
+from .kernel.base_kernel import BaseKernel
+from .mean.base_mean import BaseMean
+from .mean.means import ZeroMean
+
 from .utils.lab import pd_jitter
+from .utils.shaping import uprank_two, to_tensor
+
+"""
+At the moment the implementation of `GP` is rather messy and the inputs
+are not consistent - this is leading to troubles with implementation of
+Bayesian Optimisation.
+
+This stage of work aims to focus on improving the quality of the GP
+code such that it is not a blocker for the BO project.
+
+
+// Stage One:
+- [x] Get everything working with 'hardcoded' shapes.
+- [x] Implement additional tests to make sure these work as intended.
+
+// Stage Two:
+- [x] Add decorators to allow less constrained input (for users).
+- [x] Implement tests for decorators to ensure they work as intended.
+- [x] Add user-friendly interfaces such as __call__ and __repr__.
+"""
 
 
 class GP(nn.Module):
 
-    def __init__(self, kernel: Kernel, mean: Mean = ZeroMean(), train_noise: bool = False) -> None:
-
-        """Gaussian Process Class.
-
-        Parameters
-        ----------
-        kernel : Kernel
-            Positive-Definite Kernel for calculations.
-        train_noise : bool
-            True if training with noise, False otherwise.
-        """
+    def __init__(self,
+                 kernel: BaseKernel,
+                 mean: BaseMean = ZeroMean(),
+                 noise: bool = False) -> None:
 
         super().__init__()
 
         if not len(mean) == len(kernel):
-            raise ValueError('Need same number of means and kernels.')
+            raise ValueError('len(mean) must equal len(kernel).')
 
         self.mean = mean
         self.kernel = kernel
-        self.likelihood = GaussianLikelihood()
 
         self.x = None
         self.y = None
 
-        self.noise = self._set_noise(train_noise)
+        self.noise = self._set_noise(noise)
         self.optimiser = torch.optim.Adam
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return f'GP({str(self.mean)}, {str(self.kernel)})'
 
-    def __or__(self, other: Sequence[Any]) -> 'GP':
+    def __or__(self, other: Tuple[Tensor, Tensor]) -> 'GP':
 
         if not len(other) == 2:
             raise ValueError('Must provide (x, y)')
@@ -55,149 +66,25 @@ class GP(nn.Module):
         self.observe(*other)
         return self
 
-    @to_tensor
-    @uprank_two
-    def __call__(self, xp: Tensor) -> Normal:
+    def __call__(self, xp: Any) -> MultivariateNormal:
         return self.posterior(xp)
-
-    @property
-    def n_outputs(self) -> int:
-        return len(self.mean)
 
     @staticmethod
     def _set_noise(noise: bool) -> Tensor:
-
-        """Sets Noise to either Zero / a Parameter.
-
-        Parameters
-        ----------
-        noise : bool
-            True if noise is to be a parameter, False otherwise.
-
-        Returns
-        -------
-        Tensor
-            Parameter if noise is required, Zero otherwise.
-        """
 
         if noise:
             return nn.Parameter(torch.tensor(1.0, dtype=torch.float32))
         else:
             return torch.tensor(0.0, dtype=torch.float32)
 
-    def log_likelihood(self, stable: bool = True) -> Tensor:
-
-        """Calculates the Log Likelihood of the Observations.
-
-        Parameters
-        ----------
-        stable : bool
-            True if using autograd, False otherwise.
-
-        Returns
-        -------
-        ll : Tensor
-            Log Likelihood of the observed events.
-        """
-
-        ll = self.likelihood.log_likelihood(
-            self.kernel, self.x, unwrap(self.y), self.noise, stable
-        )
-        return ll
-
-    def optimise(self, max_iter: int = 1000, patience: int = 10) -> Tensor:
-
-        """Optimise GP Parameters.
-
-        Parameters
-        ----------
-        max_iter : int
-            Number of iterations to run.
-        patience : int
-            Patience used in EarlyStopping.
-
-        Returns
-        -------
-        loss : Tensor
-            Calculated NLL.
-        """
-
-        if self.x is None or self.y is None:
-            raise ValueError('Must set (x, y) values first.')
-
-        loss = None
-        optimiser = self.optimiser(self.parameters())
-        es = EarlyStopping(patience=patience)
-
-        for i in range(max_iter):
-
-            loss = -self.log_likelihood(stable=False)
-
-            if es(loss):
-                break
-
-            def closure():
-                optimiser.zero_grad()
-                loss.backward()
-                return loss
-
-            optimiser.step(closure)
-
-        return loss
-
-    def train(self, n_restarts: int = 10, n_iterations: int = 2000) -> Tensor:
-
-        """Train the GP Model.
-
-        Parameters
-        ----------
-        n_restarts : int
-            Number of times to restart the optimisation process.
-        n_iterations : int
-            Number of iterations to optimise for.
-
-        Returns
-        -------
-        loss : Tensor
-            Calculated NLL of the best optimisation attempt.
-        """
-
-        best_loss = None
-        best_params = deepcopy({k: v for k, v in self.named_parameters()})
-
-        for i in range(n_restarts):
-
-            for param in self.parameters():
-                torch.nn.init.uniform_(param, 0.0, 1.0)
-
-            loss = self.optimise(max_iter=n_iterations)
-
-            if best_loss is None or loss < best_loss:
-                best_loss = loss
-                best_params = deepcopy({k: v for k, v in self.named_parameters()})
-
-        for name, param in self.named_parameters():
-            param.data = best_params[name]
-
-        return best_loss
-
     @to_tensor
     @uprank_two
     def observe(self, x: Any, y: Any) -> None:
 
-        """Adds additional observations.
-
-        Parameters
-        ----------
-        x : Tensor
-            Observation features.
-        y : Tensor
-            Observations.
-        """
-
-        if not y.shape[1] == self.n_outputs:
-            raise ValueError(
-                f'shape of y ({y.shape}) does not match n_outputs.')
+        if not len(x.shape) == len(y.shape) == 2:
+            raise ValueError('Must provide 2D inputs.')
+        if not x.shape[0] == y.shape[0]:
+            raise ValueError('Must provide same number of samples for x and y.')
 
         if self.x is None and self.y is None:
             self.x = x
@@ -206,30 +93,12 @@ class GP(nn.Module):
             self.x = torch.cat([self.x, x], dim=0)
             self.y = torch.cat([self.y, y], dim=0)
 
-    def prior(self) -> MultivariateNormal:
-
-        k_xx = self.kernel.calculate(self.x, self.x)
-        mu = self.mean.calculate(self.x).flatten()
-        k_xx = pd_jitter(k_xx)
-
-        return MultivariateNormal(mu, k_xx)
-
     @to_tensor
     @uprank_two
     def posterior(self, xp: Any) -> MultivariateNormal:
 
-        """Predicts the Posterior.
-
-        Parameters
-        ----------
-        xp : Tensor
-            Features at which to predict the posterior.
-
-        Returns
-        -------
-        Normal
-            Normal distribution with calculated mu and covariance.
-        """
+        if not xp.shape[1] == self.x.shape[1]:
+            raise ValueError('xp shape does not match observed sample shapes.')
 
         k_xx = self.kernel.calculate(self.x, self.x)
         k_xxp = self.kernel.calculate(self.x, xp)
@@ -238,10 +107,75 @@ class GP(nn.Module):
 
         k_xx_inv = torch.inverse(k_xx + self.noise * torch.eye(k_xx.shape[0]))
 
-        p_mean = self.mean.calculate(xp) + k_xpx @ k_xx_inv @ unwrap(self.y)
+        p_mean = self.mean.calculate(xp) + k_xpx @ k_xx_inv @ self.y
         p_covariance = k_xpxp - k_xpx @ k_xx_inv @ k_xxp
 
         p_mean = p_mean.flatten()
         p_covariance = pd_jitter(p_covariance)
 
         return MultivariateNormal(p_mean, p_covariance)
+
+    def optimise(self, n_iterations: int = 1000) -> Tensor:
+
+        loss = None
+        optimiser = self.optimiser(self.parameters())
+
+        for i in range(n_iterations):
+
+            optimiser.zero_grad()
+
+            loss = -self.log_likelihood(grad=True).sum()
+
+            loss.backward()
+            optimiser.step()
+
+        return loss
+
+    def optimise_restarts(self, n_restarts: int = 10, n_iterations: int = 1000) -> Tensor:
+
+        losses = []
+        params = []
+
+        for i in range(n_restarts):
+
+            for param in self.parameters():
+                torch.nn.init.uniform_(param, 0.0, 2.0)
+
+            loss = self.optimise(n_iterations=n_iterations)
+
+            losses.append(loss)
+            params.append(deepcopy({k: v for k, v in self.named_parameters()}))
+
+        losses = torch.stack(losses, dim=0)
+        idx_best = losses.argmin()
+
+        for name, param in self.named_parameters():
+            param.data = params[idx_best][name]
+
+        return losses[idx_best]
+
+    def log_likelihood(self, grad: bool = False) -> Tensor:
+
+        k_xx = self.kernel.calculate(self.x, self.x)
+        k_xx = k_xx + self.noise * torch.eye(k_xx.shape[0])
+        k_xx = pd_jitter(k_xx)
+
+        if grad:
+            log_term = 0.5 * torch.log(1e-6 + torch.det(k_xx))
+            y_term = 0.5 * self.y.T @ torch.inverse(k_xx) @ self.y
+            const_term = 0.5 * len(self.x) * np.log(2 * np.pi)
+
+            ll = -y_term - log_term - const_term
+
+        else:
+            L = torch.cholesky(k_xx)
+            a0, _ = torch.lstsq(self.y, L)
+            alpha, _ = torch.lstsq(a0, L.T)
+
+            y_alpha = -0.5 * self.y * alpha.view_as(self.y)
+            trace_log = torch.trace(torch.log(L))
+            const = 0.5 * len(self.x) * np.log(2 * np.pi)
+
+            ll = y_alpha - trace_log - const
+
+        return ll
