--- gpybo/new_kernel/__init__.py
@@ -1,2 +1,0 @@
-from .kernel import *
-from .combination import SumKernel, ProductKernel
